Refactor: small cleanup
