+++ app/game/structure.go
@@ -23,6 +23,7 @@ type Game struct {
 	withSide     bool
 	side         set.Cards
 	playedCards  set.Cards
+	playedCard   card.Item
 	auctionScore auction.Score
 	phase        phase.ID
 	isToShow     bool
@@ -79,6 +80,7 @@ func (g *Game) LastPlaying() *list.List       { return &g.lastPlaying }
 func (g *Game) Phase() phase.ID               { return g.phase }
 func (g *Game) Players() team.Players         { return g.players }
 func (g *Game) PlayedCards() *set.Cards       { return &g.playedCards }
+func (g *Game) PlayedCard() card.Item         { return g.playedCard }
 func (g *Game) SideDeck() *set.Cards          { return &g.side }
 func (g *Game) IsSideToShow() bool            { return g.isToShow && g.phase == phase.InsideAuction }
 func (g *Game) SideSubset() *set.Cards        { return &g.sideSubset }
